Backend/h3-index/requirements.txt
affine==2.4.0
attrs==25.1.0
cassandra-driver==3.29.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
contextily==1.6.2
contourpy==1.3.1
cycler==0.12.1
dotenv==0.9.9
fonttools==4.56.0
geodatasets==2024.8.0
geographiclib==2.0
geomet==0.2.1.post1
geopandas==1.0.1
geopy==2.4.1
h3==4.2.1
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
mercantile==1.2.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
rasterio==1.4.3
requests==2.32.3
shapely==2.0.7
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
xyzservices==2025.1.0
